feat: Add new folder with models and routes
